Revert "feat(core): add releasesFactory to set releases list"
