+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hawk.api",
-  "version": "1.1.18",
+  "version": "1.1.17",
   "main": "index.ts",
   "license": "UNLICENSED",
   "scripts": {
