+++ src/index.ts
@@ -26,7 +26,6 @@ import PlansFactory from './models/plansFactory';
 import BusinessOperationsFactory from './models/businessOperationsFactory';
 import schema from './schema';
 import { graphqlUploadExpress } from 'graphql-upload';
-import ReleasesFactory from './models/releaseFactory';
 
 /**
  * Option to enable playground
@@ -146,16 +145,12 @@ class HawkAPI {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     const businessOperationsFactory = new BusinessOperationsFactory(mongo.databases.hawk!, dataLoaders);
 
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const releasesFactory = new ReleasesFactory(mongo.databases.events!, dataLoaders);
-
     return {
       usersFactory,
       workspacesFactory,
       projectsFactory,
       plansFactory,
       businessOperationsFactory,
-      releasesFactory,
     };
   }
