+++ src/typeDefs/index.ts
@@ -17,7 +17,6 @@ import workspaceMutations from './workspaceMutations';
 import chart from './chart';
 import plans from './plans';
 import seed from './seed';
-import release from './release';
 import isE2E from '../utils/isE2E';
 
 const rootSchema = gql`
@@ -101,7 +100,6 @@ const typeDefinitions = [
   workspaceMutations,
   chart,
   plans,
-  release,
   projectEventGroupingPattern,
   projectEventGroupingPatternMutations,
 ];
